… examples/common.cpp
… examples/common.h
… examples/embedding/embedding.cpp
… examples/main/README.md
… examples/main/main.cpp
… examples/perplexity/perplexity.cpp
… examples/quantize/quantize.cpp
… examples/server/server.cpp
… examples/simple/simple.cpp
… llama.cpp
… llama.h