2S Xeon E5-2690v1

Before:
./main -n 1024 -m models/7B/ggml-model-q4_0.bin --ignore-eos -p "Building a website can be done in 10 simple steps:" -t 32
...
llama_print_timings:      sample time =   796.65 ms /  1024 runs   (    0.78 ms per token)
llama_print_timings: prompt eval time = 44892.59 ms /   785 tokens (   57.19 ms per token)
llama_print_timings:        eval time = 178337.97 ms /  1020 runs   (  174.84 ms per token)

\# echo 3 > /proc/sys/vm/drop_caches

After:
./main -n 1024 -m models/7B/ggml-model-q4_0.bin --ignore-eos -p "Building a website can be done in 10 simple steps:" -t 32
...
llama_print_timings:      sample time =   795.63 ms /  1024 runs   (    0.78 ms per token)
llama_print_timings: prompt eval time = 46996.62 ms /   785 tokens (   59.87 ms per token)
llama_print_timings:        eval time = 101232.21 ms /  1020 runs   (   99.25 ms per token)