@zrm 2023-05-21
Not sure if this one is alright (9d058c2096b9f1f300e1ee16f5740a6a0a342917). On machines with high thread synchronization overhead, distributing GGML_TASK_FINALIZE to the thread pool adds about 10% to eval time, so only do it for ops where finalize actually does something. But am I correct that this is currently none of them?