@ggerganov 2023-05-24
Here are some results on a dual socket `Intel(R) Xeon(R) Gold 6154 CPU @ 3.00GHz`. Each one has 18 cores

`master`

```bash
# 8 threads
make -j && ./main -m models/7B/ggml-model-q4_0.bin -p "I believe the meaning of life is" --no-mmap -n 32 -t 8 -s 6

llama_print_timings:        load time =  2649.90 ms
llama_print_timings:      sample time =    14.44 ms /    26 runs   (    0.56 ms per token)
llama_print_timings: prompt eval time =   488.38 ms /     8 tokens (   61.05 ms per token)
llama_print_timings:        eval time =  3450.39 ms /    25 runs   (  138.02 ms per token)
llama_print_timings:       total time =  6120.78 ms

# 12 threads
make -j && ./main -m models/7B/ggml-model-q4_0.bin -p "I believe the meaning of life is" --no-mmap -n 32 -t 12 -s 6

llama_print_timings:        load time =  2558.62 ms
llama_print_timings:      sample time =    13.90 ms /    26 runs   (    0.53 ms per token)
llama_print_timings: prompt eval time =   335.17 ms /     8 tokens (   41.90 ms per token)
llama_print_timings:        eval time =  2190.39 ms /    25 runs   (   87.62 ms per token)
llama_print_timings:       total time =  4768.89 ms

# 16 threads
make -j && ./main -m models/7B/ggml-model-q4_0.bin -p "I believe the meaning of life is" --no-mmap -n 32 -t 16 -s 6

llama_print_timings:        load time =  2422.32 ms
llama_print_timings:      sample time =    13.93 ms /    26 runs   (    0.54 ms per token)
llama_print_timings: prompt eval time =   264.24 ms /     8 tokens (   33.03 ms per token)
llama_print_timings:        eval time =  2008.39 ms /    25 runs   (   80.34 ms per token)
llama_print_timings:       total time =  4450.76 ms

# 18 threads
make -j && ./main -m models/7B/ggml-model-q4_0.bin -p "I believe the meaning of life is" --no-mmap -n 32 -t 18 -s 6

llama_print_timings:        load time =  2467.76 ms
llama_print_timings:      sample time =    14.18 ms /    26 runs   (    0.55 ms per token)
llama_print_timings: prompt eval time =   306.80 ms /     8 tokens (   38.35 ms per token)
llama_print_timings:        eval time =  1668.66 ms /    25 runs   (   66.75 ms per token)
llama_print_timings:       total time =  4156.82 ms

# 36 threads
make -j && ./main -m models/7B/ggml-model-q4_0.bin -p "I believe the meaning of life is" --no-mmap -n 32 -t 36 -s 6


llama_print_timings:        load time =  2440.28 ms
llama_print_timings:      sample time =    13.64 ms /    26 runs   (    0.52 ms per token)
llama_print_timings: prompt eval time =   283.70 ms /     8 tokens (   35.46 ms per token)
llama_print_timings:        eval time =  2977.92 ms /    25 runs   (  119.12 ms per token)
llama_print_timings:       total time =  5438.13 ms
```

`zrm:numa`

```bash
# 8 threads
make -j && ./main -m models/7B/ggml-model-q4_0.bin -p "I believe the meaning of life is" --no-mmap -n 32 -t 8 -s 6

llama_print_timings:        load time =  2784.40 ms
llama_print_timings:      sample time =    12.51 ms /    26 runs   (    0.48 ms per token)
llama_print_timings: prompt eval time =   626.98 ms /     8 tokens (   78.37 ms per token)
llama_print_timings:        eval time =  3395.67 ms /    25 runs   (  135.83 ms per token)
llama_print_timings:       total time =  6198.56 ms

# 12 threads
make -j && ./main -m models/7B/ggml-model-q4_0.bin -p "I believe the meaning of life is" --no-mmap -n 32 -t 12 -s 6

llama_print_timings:        load time =  2611.76 ms
llama_print_timings:      sample time =    12.17 ms /    26 runs   (    0.47 ms per token)
llama_print_timings: prompt eval time =   407.14 ms /     8 tokens (   50.89 ms per token)
llama_print_timings:        eval time =  2819.72 ms /    25 runs   (  112.79 ms per token)
llama_print_timings:       total time =  5449.82 ms

# 16 threads
make -j && ./main -m models/7B/ggml-model-q4_0.bin -p "I believe the meaning of life is" --no-mmap -n 32 -t 16 -s 6

llama_print_timings:        load time =  2542.99 ms
llama_print_timings:      sample time =    11.76 ms /    26 runs   (    0.45 ms per token)
llama_print_timings: prompt eval time =   335.91 ms /     8 tokens (   41.99 ms per token)
llama_print_timings:        eval time =  2551.17 ms /    25 runs   (  102.05 ms per token)
llama_print_timings:       total time =  5111.75 ms

# 18 threads
make -j && ./main -m models/7B/ggml-model-q4_0.bin -p "I believe the meaning of life is" --no-mmap -n 32 -t 18 -s 6

llama_print_timings:        load time =  2500.30 ms
llama_print_timings:      sample time =    11.47 ms /    26 runs   (    0.44 ms per token)
llama_print_timings: prompt eval time =   300.49 ms /     8 tokens (   37.56 ms per token)
llama_print_timings:        eval time =  2538.30 ms /    25 runs   (  101.53 ms per token)
llama_print_timings:       total time =  5055.94 ms

# 36 threads
make -j && ./main -m models/7B/ggml-model-q4_0.bin -p "I believe the meaning of life is" --no-mmap -n 32 -t 36 -s 6

llama_print_timings:        load time =  2448.93 ms
llama_print_timings:      sample time =    11.44 ms /    26 runs   (    0.44 ms per token)
llama_print_timings: prompt eval time =   253.50 ms /     8 tokens (   31.69 ms per token)
llama_print_timings:        eval time =  2497.20 ms /    25 runs   (   99.89 ms per token)
llama_print_timings:       total time =  4963.27 ms
```

I don't observe the speed-up from the `zrm:numa` branch.
I suspect that this has something to do with `mmap`.

@zrm has done their tests without the `--no-mmap` option, so this means by default `mmap` is used.
On my system, if I use `mmap`, there is a significant slowdown - here is `master` using 18 threads:

```bash
make -j && ./main -m models/7B/ggml-model-q4_0.bin -p "I believe the meaning of life is" -n 32 -t 18 -s 6

llama_print_timings:        load time =   434.70 ms
llama_print_timings:      sample time =    14.18 ms /    26 runs   (    0.55 ms per token)
llama_print_timings: prompt eval time =   259.91 ms /     8 tokens (   32.49 ms per token)
llama_print_timings:        eval time =  3285.42 ms /    25 runs   (  131.42 ms per token)
llama_print_timings:       total time =  3740.55 ms
```

I recommend to re-do the analysis with always using `--no-mmap` and see if there is benefit from NUMA in that case