@rankaiyx 2023-06-17
If the number of threads is odd, the result exceeds the maximum number of nodes in numa.

One solution is:
int node_num = thread_n / ((float)n_threads / ggml_numa.n_nodes);

A better solution (which can improve performance by 20%) is:
int node_num = thread_n % ggml_numa.n_nodes;

Only 13B and 65B were tested.
13B performance improvement of 20% (from 122ms to 105ms).
The performance of 65B has not improved (from 439ms to 440ms).